Migrate back to JS - We are moving back to JS do to some issues with our config.json file - When the CLI makes changes to the file it requiers the use to rebuild the project breaking the UX - JS does not have the same issues. When we have more time we can consider migrating to TS again

+++ src/protocol/functions/buyWaterCooler.js
@@ -46,7 +46,7 @@ export default async () => {
 
     tx.setGasBudget(config.gasBudgetAmount);
 
-    const [coin] = tx.splitCoins(tx.gas, [price as bigint]);
+    const [coin] = tx.splitCoins(tx.gas, [price]);
 
     const coolerFactoryId = getCoolerFactoryId();
